FSM MainServer Broadcast 구조체 전달하긴했는데 ..

+++ IOCP_Broadcast/BlueNpc.h
@@ -62,7 +62,7 @@ class BlueNpc
 	struct minion
 	{
 		int my_index = 0;
-		float npc_pos_[3];
+		int npc_pos_[3];
 		int npc_pos_dir_ = 5;
 		int team_color = 0;
 		int npc_speed = 1;
